Changes regarding review Add database name to SP. Move generation to INventoryDataManager Fix test
Issue: IBP-3146
Reviewer: Clary
