+++ src/test/java/org/generationcp/middleware/service/impl/inventory/LotIDGeneratorTest.java
@@ -1,10 +1,12 @@
 package org.generationcp.middleware.service.impl.inventory;
 
 import org.apache.commons.lang3.RandomStringUtils;
+import org.generationcp.middleware.manager.InventoryDataManagerImpl;
 import org.generationcp.middleware.pojos.ims.Lot;
 import org.generationcp.middleware.pojos.workbench.CropType;
 import org.junit.Before;
 import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.Arrays;
 
@@ -13,13 +15,16 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-public class LotIDGeneratorImplTest {
+public class LotIDGeneratorTest {
 
 	public static final String UUID_REGEX = "[0-9a-f]{8}-[0-9a-f]{4}-[4][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}";
 	private static final int CROP_PREFIX_LENGTH = 10;
-	private static final String SUFFIX_REGEX = "[a-zA-Z0-9]{" + LotIDGeneratorImpl.SUFFIX_LENGTH + "}";
+	private static final String SUFFIX_REGEX = "[a-zA-Z0-9]{" + InventoryDataManagerImpl.SUFFIX_LENGTH + "}";
 
-	private final LotIDGeneratorImpl lotIDGenerator = new LotIDGeneratorImpl();
+
+	@Autowired
+	private InventoryDataManagerImpl inventoryDataManager;
+
 	private CropType crop;
 
 	@Before
@@ -27,6 +32,7 @@ public class LotIDGeneratorImplTest {
 		this.crop = new CropType();
 		this.crop.setPlotCodePrefix(RandomStringUtils.randomAlphanumeric(CROP_PREFIX_LENGTH));
 		this.crop.setUseUUID(true);
+		this.inventoryDataManager = new InventoryDataManagerImpl();
 	}
 
 	@Test
@@ -34,14 +40,14 @@ public class LotIDGeneratorImplTest {
 		final Lot lot = new Lot();
 		final String existingLotId = RandomStringUtils.randomAlphanumeric(20);
 		lot.setLotUuId(existingLotId);
-		this.lotIDGenerator.generateLotIds(this.crop, Arrays.asList(lot));
+		this.inventoryDataManager.generateLotIds(this.crop, Arrays.asList(lot));
 		assertEquals(existingLotId, lot.getLotUuId());
 	}
 
 	@Test
 	public void testGenerateLotIds_UseUUID() {
 		final Lot lot = new Lot();
-		this.lotIDGenerator.generateLotIds(this.crop, Arrays.asList(lot));
+		this.inventoryDataManager.generateLotIds(this.crop, Arrays.asList(lot));
 		assertNotNull(lot.getLotUuId());
 		assertTrue(lot.getLotUuId().matches(UUID_REGEX));
 	}
@@ -50,11 +56,11 @@ public class LotIDGeneratorImplTest {
 	public void testGenerateLotIds_UseCustomID() {
 		this.crop.setUseUUID(false);
 		final Lot lot = new Lot();
-		this.lotIDGenerator.generateLotIds(this.crop, Arrays.asList(lot));
+		this.inventoryDataManager.generateLotIds(this.crop, Arrays.asList(lot));
 		final String lotId = lot.getLotUuId();
 		assertNotNull(lotId);
 		assertFalse(lotId.matches(UUID_REGEX));
-		assertEquals(this.crop.getPlotCodePrefix() + LotIDGeneratorImpl.MID_STRING, lotId.substring(0, CROP_PREFIX_LENGTH + 1));
+		assertEquals(this.crop.getPlotCodePrefix() + InventoryDataManagerImpl.MID_STRING, lotId.substring(0, CROP_PREFIX_LENGTH + 1));
 		final String suffix = lotId.substring(CROP_PREFIX_LENGTH + 1, lotId.length());
 		assertTrue(suffix.matches(SUFFIX_REGEX));
 	}
